Moved Context object to new subfolder
