+++ ClientApp/src/Components/Auth/AuthenticatedRoute/AuthenticatedRoute.tsx
@@ -1,6 +1,6 @@
 ﻿import * as React from "react";
 import { Route, Redirect, RouteProps } from "react-router-dom";
-import { AuthContext } from "../../../Context/AuthContext";
+import { AuthContext } from "../../../Context/AuthContext/AuthContext";
 import { routeDefinitions } from "../../../Router/routeDefinitions";
 const AuthenticatedRoute: React.FC<RouteProps> = ({
   path,
